Step3: Implement `let*` function using `def_fn()`
`let*` function pushes a new context into the env with an iterable (list
or vector) to initialize the new environment. Then, second agrument is
evaluated in the new environment, and finally the new context is popped
off the stack.

Implemented helper method on dyn MalType to access iterables' backing
arrays, if given Maltype is not iterable then MalType::TypeError is
raised.