Step7: Implement quoting and core list functions
Quoting is done using `quote`, and `quasiquote` functions where `quote`
can be used to "escape" and interpret its arguments as AST rather than
values. `quasiquote` on the other hand can be used to interpret quoted
AST, and produce values.

Core list functions added are `cons` which appends its first arg to list
in the second argument. `concat` takes in number of lists, and
concatenates them into a single list. `vec` function takes in a single
argument as list, and returns it as list.