Wrap `is_macro` field of MalClojure
This change allows a clojure to become a macro after instantiation.
Previous to this change, a clojure cannot be made into a macro if it's
not instantiated with `defmacro!` since there would be multiple
references to clojure (at least one from the AST, and one from env), and
`Rc<_>` prevents multiple owning (or mutalbe) referances.